Update #2
In this update, I added an Id tag to the main hero image, and began stylizing the main card. I added the root folder containing the colors and will add fonts and font sizing and weights later. This was an early roughing-in for this project, and will get more in depth in the next update.